requirements.txt
FastAPI
langchain-openai
langchain-google-genai
langchain-community
python-dotenv